backend_v2/ml_training/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "skincare-ml-training"
version = "0.1.0"
description = "PyTorch training pipeline for the skincare multi-label skin condition classifier"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "timm>=1.0.9",
    "numpy>=1.26.0",
    "pillow>=10.4.0",
    "opencv-python-headless>=4.10.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.66.0",
    "tensorboard>=2.17.0",
    "matplotlib>=3.9.0",
    "onnx>=1.17.0",
    "onnxruntime>=1.19.0",
    "grad-cam>=1.5.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.6.0",
]

[project.scripts]
skin-train = "skin_training.train.loop:main"
skin-precompute = "skin_training.data.precompute:main"
skin-export = "skin_training.export.to_onnx:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["skin_training*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
